fix: fix work with absolute paths and display of messages.
