+++ src/commands/hash.js
@@ -1,18 +1,15 @@
 import path from 'node:path';
 import crypto from 'crypto';
 import { createReadStream } from 'fs';
-import { getCurrentPath } from '../currentPath.js'
+import { checkArgumentsCount, error, getCurrentPath, currentPathMessage } from '../helpers/utils.js';
 
 export const hash = async (args) => {
-    if(args.length !== 1) {
-        console.log('Invalid input');
-        return;
-    }
+    if (checkArgumentsCount(args, 1)) return;
 
     const fullFilePath = path.isAbsolute(args[0]) ? args[0] : path.join(getCurrentPath(), args[0]);
     const hash = crypto.createHash('sha256');
 
-    await new Promise((resolve, reject) => {
+    await new Promise((resolve) => {
         const readStream = createReadStream(fullFilePath);
 
         readStream.on('data', function(fileHash) {
@@ -20,13 +17,15 @@ export const hash = async (args) => {
         });
 
     
-        readStream.on('error', () => {
-            reject(new Error(`Operation failed!`));
+        readStream.on('error', (err) => {
+            error(err);
+            currentPathMessage();
         });
     
         readStream.on('end', () => {
             const fileHash = hash.digest('hex');
             console.log(`Hash of ${ fullFilePath }: ` + fileHash);
+            currentPathMessage();
             resolve();
         });
     });
